requirements.txt
Flask
mysql-connector-python
pandas
numpy
scikit-learn
joblib
